Change Databricks 310 shim to be 311 because they changed the spark.version returned
