+++ pom.xml
@@ -133,7 +133,7 @@
             </properties>
         </profile>
         <profile>
-            <id>databricks310</id>
+            <id>databricks311</id>
             <properties>
                 <rat.consoleOutput>true</rat.consoleOutput>
             </properties>
@@ -145,9 +145,9 @@
             </properties>
         </profile>
         <profile>
-            <id>spark310dbtests</id>
+            <id>spark311dbtests</id>
             <properties>
-                <spark.test.version>${spark310db.version}</spark.test.version>
+                <spark.test.version>${spark311db.version}</spark.test.version>
             </properties>
         </profile>
         <profile>
@@ -238,7 +238,7 @@
         <spark301emr.version>3.0.1-amzn</spark301emr.version>
         <spark302.version>3.0.2</spark302.version>
         <spark303.version>3.0.3-SNAPSHOT</spark303.version>
-        <spark310db.version>3.1.0-databricks</spark310db.version>
+        <spark311db.version>3.1.1-databricks</spark311db.version>
         <spark311.version>3.1.1</spark311.version>
         <spark312.version>3.1.2-SNAPSHOT</spark312.version>
         <spark320.version>3.2.0-SNAPSHOT</spark320.version>
